+++ src/Array.ts
@@ -1,68 +1,82 @@
 import assert from "assert";
 import Base from "./Any";
-import { number, TYPE } from "./utils";
+import ObjectType from "./Object";
+import { mergeErrors, NULL, number, TYPE } from "./utils";
 
 const { isNumber } = number;
 
-class Type<T = any> extends Base<T[]> {
+class Type<T = any, P extends Base = Base> extends Base<T[], { items?: P }> {
   protected static type = TYPE.ARRAY;
 
   public min(num: number) {
-    assert(isNumber(num), "'num' must be a number");
-    assert(num >= 0, "'num' must be a positive number");
+    assert(isNumber(num) && num >= 0, "Expected num to be a positive number");
 
     return this._pipe(value => ({
       value,
-      errors: value.length < num ? `Must be at least ${num} item(s)` : {},
+      errors:
+        value.length < num
+          ? `Expected to contain at least ${num} item(s)`
+          : NULL,
     }));
   }
 
   public max(num: number) {
-    assert(isNumber(num), "'num' must be a number");
-    assert(num >= 0, "'num' must be a positive number");
+    assert(isNumber(num) && num >= 0, "Expected num to be a positive number");
 
     return this._pipe(value => ({
       value,
-      errors: value.length > num ? `Must be at most ${num} item(s)` : {},
+      errors:
+        value.length > num
+          ? `Expected to contain at most ${num} item(s)`
+          : NULL,
     }));
   }
 
   public length(num: number) {
-    assert(isNumber(num), "'num' must be a number");
-    assert(num >= 0, "'num' must be a positive number");
+    assert(isNumber(num) && num >= 0, "Expected num to be a positive number");
 
     return this._pipe(value => ({
       value,
-      errors: value.length !== num ? `Must be exactly ${num} item(s)` : {},
+      errors:
+        value.length !== num
+          ? `Expected to contain exactly ${num} item(s)`
+          : NULL,
     }));
   }
 
-  public items(type: Base) {
+  public items(type: P) {
     assert(
       Base.isType(type),
-      `Expected 'type' to be a 'TypeAny' instance, got '${typeof type}' instead`,
+      `Expected type to be a valid schema type, got ${typeof type}`,
     );
 
-    return this._pipe((value, path) =>
+    this.details.items = type;
+
+    return this._pipe(value =>
       value.reduce(
         (prev, item, index) => {
-          const result = (type as any)._validate(`${path}[${index}]`, item);
+          const result = type.validate(item);
 
           prev.value.push(result.value);
 
-          prev.errors = Object.assign({}, prev.errors, result.errors);
+          prev.errors = mergeErrors<T[]>(prev.errors, {
+            [`[${index}]`]: result.errors,
+          } as any);
 
           return prev;
         },
-        { value: [] as any[], errors: {} },
+        {
+          value: [] as any[],
+          errors: NULL as Base.ValidationResult<T[]>["errors"],
+        },
       ),
     );
   }
 
   protected _base(value: any) {
-    if (Array.isArray(value)) return null;
+    if (Array.isArray(value)) return NULL;
 
-    return "Must be an array";
+    return "Expected to be an array";
   }
 }